fix: Moves shared-video middleware to use CONFERENCE_JOIN_IN_PROGRESS.
